Load one locale
